pyproject.toml
[project]
name = "deepsdf-for-bim-geometry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ifcopenshell>=0.8.3.post2",
    "numpy>=2.3.2",
    "open3d>=0.19.0",
    "point-cloud-utils>=0.34.0",
    "pyrender>=0.1.45",
    "pyvista>=0.46.3",
    "pyyaml>=6.0.2",
    "questionary>=2.1.0",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "rtree>=1.4.1",
    "scikit-image>=0.25.2",
    "streamlit>=1.51.0",
    "tensorboard>=2.20.0",
    "torch",
    "torchaudio",
    "torchvision",
    "tqdm-joblib>=0.0.5",
    "trimesh>=4.7.3",
    "unzip>=1.0.0",
    "vtk>=9.5.2",
]

[tool.uv.sources]
torch = { url = "https://download.pytorch.org/whl/cu124/torch-2.4.0%2Bcu124-cp312-cp312-linux_x86_64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/cu124/torchvision-0.19.0%2Bcu124-cp312-cp312-linux_x86_64.whl" }
torchaudio = { url = "https://download.pytorch.org/whl/cu124/torchaudio-2.4.0%2Bcu124-cp312-cp312-linux_x86_64.whl" }

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
